Move PrestoAction related classes into a separate package
