+++ presto-verifier/src/main/java/com/facebook/presto/verifier/PrestoVerifyCommand.java
@@ -16,9 +16,9 @@ package com.facebook.presto.verifier;
 import com.facebook.presto.sql.parser.SqlParserOptions;
 import com.facebook.presto.sql.tree.Property;
 import com.facebook.presto.verifier.framework.AbstractVerifyCommand;
-import com.facebook.presto.verifier.framework.PrestoExceptionClassifier;
 import com.facebook.presto.verifier.framework.SourceQuery;
-import com.facebook.presto.verifier.framework.SqlExceptionClassifier;
+import com.facebook.presto.verifier.prestoaction.PrestoExceptionClassifier;
+import com.facebook.presto.verifier.prestoaction.SqlExceptionClassifier;
 import com.facebook.presto.verifier.resolver.ExceededGlobalMemoryLimitFailureResolver;
 import com.facebook.presto.verifier.resolver.ExceededTimeLimitFailureResolver;
 import com.facebook.presto.verifier.resolver.FailureResolver;
