+++ presto-verifier/src/main/java/com/facebook/presto/verifier/framework/QueryRewriter.java
@@ -26,7 +26,8 @@ import com.facebook.presto.sql.tree.QualifiedName;
 import com.facebook.presto.sql.tree.Query;
 import com.facebook.presto.sql.tree.QuerySpecification;
 import com.facebook.presto.sql.tree.Statement;
-import com.facebook.presto.verifier.framework.PrestoAction.ResultSetConverter;
+import com.facebook.presto.verifier.prestoaction.PrestoAction;
+import com.facebook.presto.verifier.prestoaction.PrestoAction.ResultSetConverter;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.intellij.lang.annotations.Language;
