+++ presto-verifier/src/main/java/com/facebook/presto/verifier/prestoaction/PrestoActionFactory.java
@@ -11,14 +11,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.verifier.framework;
+package com.facebook.presto.verifier.prestoaction;
 
-import com.facebook.presto.jdbc.QueryStats;
+import com.facebook.presto.verifier.framework.QueryConfiguration;
+import com.facebook.presto.verifier.framework.VerificationContext;
 
-import java.sql.SQLException;
-import java.util.Optional;
-
-public interface SqlExceptionClassifier
+public interface PrestoActionFactory
 {
-    QueryException createException(QueryStage queryStage, Optional<QueryStats> queryStats, SQLException cause);
+    PrestoAction create(
+            QueryConfiguration controlConfiguration,
+            QueryConfiguration testConfiguration,
+            VerificationContext verificationContext);
 }
